# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VerdictIQ - Court Judgment Intelligence System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases/{id}/directives:
    get:
      operationId: listDirectives
      tags: [directives]
      summary: List all extracted directives for a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: classification
          in: query
          schema:
            type: string
            enum: [mandatory, advisory, all]
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of directives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Directive"
  /cases/{id}/directives/{directiveId}:
    get:
      operationId: getDirective
      tags: [directives]
      summary: Get a specific directive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: directiveId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Directive details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Directive"
  /cases/{id}/directives/{directiveId}/verify:
    post:
      operationId: verifyDirective
      tags: [verification]
      summary: Approve, edit, or reject an extracted directive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: directiveId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyDirectiveBody"
      responses:
        "200":
          description: Verification recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Directive"
...
  /cases/{id}/comments:
    get:
      operationId: listCaseComments
      tags: [cases]
      summary: List comments for a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseComment"
    post:
      operationId: createCaseComment
      tags: [cases]
      summary: Add a comment to a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseComment"
...
  /users/{clerkId}/role:
    patch:
      operationId: updateUserRole
      tags: [auth]
      summary: Update a user role (admin only)
      parameters:
        - name: clerkId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    Directive:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        judgmentId:
          type: integer
        type:
          type: string
          enum: [compliance_order, stay, direction, limitation_period, appeal, observation, other]
        classification:
          type: string
          enum: [mandatory, advisory, unknown]
        sourceText:
          type: string
        pageNumber:
          type: integer
        paragraphRef:
          type: string
          nullable: true
        deadline:
          type: string
          format: date
          nullable: true
        deadlineInferred:
          type: boolean
        deadlineSource:
          type: string
          nullable: true
        responsibleDepartment:
          type: string
          nullable: true
        actionRequired:
          type: string
        isNovel:
          type: boolean
        confidenceScore:
          type: number
        verificationStatus:
          type: string
          enum: [pending, approved, edited, rejected]
        reviewerName:
          type: string
          nullable: true
        reviewerCorrection:
          type: string
          nullable: true
        reviewerReason:
          type: string
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - caseId
        - judgmentId
        - type
        - classification
        - sourceText
        - pageNumber
        - actionRequired
        - isNovel
        - confidenceScore
        - verificationStatus
        - deadlineInferred
        - createdAt
    VerifyDirectiveBody:
      type: object
      properties:
        decision:
          type: string
          enum: [approved, edited, rejected]
        correctedValue:
          type: string
          nullable: true
        correctedClassification:
          type: string
          enum: [mandatory, advisory]
          nullable: true
        correctedDeadline:
          type: string
          format: date
          nullable: true
        correctedDepartment:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
        reviewerName:
          type: string
      required: [decision, reviewerName]
...
    UserRole:
      type: string
      enum: [admin, reviewer, viewer]
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        fullName:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt:
          type: string
          format: date-time
      required: [id, clerkId, email, role, createdAt]
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
      required: [role]
...
    CaseComment:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        authorName:
          type: string
        authorRole:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, caseId, authorName, authorRole, content, createdAt]
    CreateCaseCommentBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
...